Fixed cache bug, added new feature indicating squares others have clicked
cache bug, the cache was saving before a user got created, so if they refreshed or left the page before creating a user they would not get one when they tried the next visit to the site. Clearing the cache would fix it. Fixed it with this update.

New feature now animates squares when at least 2 people have clicked on it, indicating you "missed" one basically