Fixed bug with joining multiple rooms
Before if you opened two tabs with two different rooms, the second tab would use the cache from the first room. Instead, the program now checks for the room name and creates new cache if it's different. This means if you actually wanted to open multiple rooms at once, you'd need incognito browsers or a different browser which is the intended functionality for this update.