Fixed potential cache error
In the situation where new information for bingo squares is added to the database within 24 hours of a user playing. That user tries to view a user's board who has not played (hence loading fresh data not from cache). This update fixes that case where there would be an empty square on their board.